searchCircles 関連の処理をサークルのモデルファイルではなく、PDFを生成するモデルファイルに処理させるようにした。(Issue #41)
