+++ app/Models/Circle.php
@@ -73,35 +73,11 @@ class Circle extends Model {
     }
 
     /*
-     * searchCircles テーブルからデータを取得する。
-     */
-    public function getSearch($id) {
-        $searchCircles = DB::table('searchCircles')
-            ->where('event_id', $id)
-            ->get();
-
-        $getSearch = end($searchCircles);
-
-        return $getSearch;
-    }
-
-    /*
      * searchCircles テーブルのデータを取得して削除する。
      */
     public function searchCirclesDelete() {
         DB::table('searchCircles')
             ->delete();
-    }
-
-    /*
-     * 検索結果から PDF を出力するために circles テーブルから検索結果を返す
-     */
-    public function searchCircles($id, $content, $text) {
-        $searchCircles = Circle::where('event_id', $id)
-            ->where($content, 'like', '%'. $text .'%')
-            ->get();
-
-        return $searchCircles;
     }
 
     /*
